requirements.txt
flask==0.11.1
flask-sqlalchemy==2.1
pyserial==3.2.1
